chore: refactor get block methods to util
